Title: First lines of JWT & route protection && new folder
body: This push was focused on beginning to implement a
'signup' route. To do so, I realized that I was going to
have to save user data, which means that I would need to
practice secure data handling, which means I would need to
hash any sensitive data being saved. I implemented a
pre-save middleware in the userModels file to accomplish
this. I also wanted to send a JWT alongside the response
that is delivered upon successful account creation, which led
to me developing a middleware (protectRoute) that will verify
JWT's and allow contextual authorization. This said
middleware will also ensure that that a user's password has
NOT been changed since the issuing of the token

Token creation has NOT been developed yet.

Note: This project is heavily commented for my own future
reference, but also with the intention of detailing my own
thought process and intentions with clarity and precision.

S/O to my future self, I am impressed. As expected.