title: Creating AppError (custom error class)
body: I spent a few hours studying and focusing on one task.
The AppError class. It's  important to return accurate and clear
error messages, and this new class will allow me to do that. It is
also quite easy to use (and was easy to write), which makes
something like AppError even more important! Very little time was
actuaLLy spent writing this 6-line utility file, most of my time
was spent asking 'why' and 'how.