title: Modularizing JSON responses & error handling continued
body: This push was all about, for the most part, modularizing
the json responses, status codes and statuses themselves. I created
a new helper function called sendJsonRes that acomplishes this.
I felt that this was of extreme importance to acomplish early for
two reason. First, my custom error handling class was unable to
be thoroughly completed without the the addition of sending a json
response. I understand darn tootin well that I am going to be
writing countless json responses, which was the second reason that
pushed me into writing this helper function to modularize json
responses so early. I used this helper function immediately in
getAllUsers controller's function, but have yet to implement it
else where. That will soon be my main focus, putting these helpers
to great use! :)