title: Further development to protectRoute middleware
body: This commit was focused on further developing the protectRoute middleware. Specifically,
I need it to reliably, consistently, and accurately verify user credentials in order to
actually implement the route that I wish to develop next, which is the deleteUser route. Part
of the development that was done to the protectRoute middleware was focused on catching errors
that occur whenever the .verify method is called on JWT tokens (to ensure that all JWTs are
valid). To do this, I wrapped it in promisify, which allowed me to attach a .catch block
in which I returned an instance of my helper function AppError. This ensures whenever
an invalid JWT token is sent, a 401 status code is sent back, and access to further activity
of the protected route is denied.

Immediately afterwards, I wanted to
ensure that the properties that are actually needed for the rest of the function were present in
the results of the fulfilled promise (id & iat). I tried to be specific in this
verification to provide as little room as possible for errors to appear and to reduce the
possibility of misuse. Only if those properties are present, do I properly destructure them for
further use in the function. I still need to ensure that the ID obtained from the decoded JWT
payload is indeed associated with a user in the Users collection of MongoDB.
This is extremely trivial, will be implemented soon!